Removed excluding last played maps, added CreateExcludeList function, changed required maximum or minimum player counts for Breach_CV, Iso, and temple
Removed exclusion of the last played maps. Due to the dying state of the game, it would be recommended to give players the choice to extend play on the current map, or to only exclude the last n played maps (where n is a small non-negative integer, preferably 1 or 2. ).  Certain maps, such as chem labs or spaceship, bring players in when player counts are low and are enjoyable for long sessions. 

Added CreateExcludeList function, which returns a table of excluded maps and a string stating the size of the excluded maps. 

Min player count for Breach_cv increased from 5 to 12 as the map is very large. Max player counts for iso and temple have been increased from 15 to 18 as the maps are robust to large player counts.